+++ src/numberto/converters.clj
@@ -80,3 +80,4 @@
   (-> num
       (BigInteger. from-radix)
       (.toString to-radix)))
+